FIX: Multi-line search text should be passed as a joined single argument, not one search per line
For opening files, it makes sense to assume line-based input (i.e. one file per line), but when doing web searches, we'd rather have the whole text passed as-is. Many error messages and stack traces span multiple lines!
This gets rid of the xargs, but we can also drop the chdir to the pane's CWD; unlike filespecs, the web search argument is not dependent on the current path (and the search command most likely is found through PATH).